in working state after refactoring
